Simple Loop and Add-Multiply optimization
A common pattern is to build a multiplication loop. Something like

  ++[>++<-]

is essentially equivalent to

  cell[0] = 2;
  while(cell[0]) {
    cell[1] += 2;
    cell[0]--;
  }

which is further equivalent to

  cell[1] += cell[0] * 2;
  cell[0] = 0;

We can provide this by first identifying that the loop is "simple,"
the pointer always ends where it started and the value is only
impacted by arithmetic. Then we look for simple loops that only
feature adds/subtracts.

This patch introduced some pseudo IR nodes, MovImm sets a memory
value, SimpleLoop is a Loop where the final delta is implied, and
AddMul to add the multiplied value.